ag-ui/gpt-researcher/requirements.txt
ag2[openai,ag-ui]>=0.11.2
gpt-researcher
fastapi
uvicorn[standard]
python-dotenv
